refactor: likes 필드 제거
- ArticleContent 엔티티에서 likes 필드를 제거하고, 좋아요 수를 동적으로 계산하도록 변경
- 관련 쿼리와 DTO 로직 수정

+++ src/main/java/com/moaguide/dto/RelatedContentDto.java
@@ -13,5 +13,5 @@ public class RelatedContentDto {
   private String imgLink;     // 대표 이미지 URL
   private Timestamp createdAt; // 작성일
   private Integer views;      // 조회수
-  private Integer likes;      // 좋아요 수
+  private Long likes;      // 좋아요 수
 }
